continuously clear old versions found
